Revert "feat(objects): add object schema factory"
