+++ speckle/resources/objects.py
@@ -4,8 +4,6 @@ from speckle.base.resource import ResourceBase
 from pydantic import BaseModel, validator
 from typing import List, Optional
 from speckle.base.resource import ResourceBase, ResourceBaseSchema
-from speckle import objects
-
 
 NAME = 'objects'
 METHODS = ['list', 'get', 'update', 'create'
@@ -13,15 +11,28 @@ METHODS = ['list', 'get', 'update', 'create'
            'get_bulk', 'set_properties']
 
 
-class SpeckleObject:
+class SpeckleObject(ResourceBaseSchema):
+    type: Optional[str]
+    name: Optional[str]
+    geometryHash: Optional[str]  # Is immediately replaced anyways
+    hash: Optional[str]  # Is immediately replaced anyways
+    applicationId: Optional[str]
+    properties: dict = {}
+    partOf: Optional[List[str]]
+    parent: Optional[List[str]]
+    children: Optional[List[str]]
+    ancestors: Optional[List[str]]
 
-    def parse_obj(self, data):
-        try:
-            assert 'type' in data, 'missing key type'
-            object_schema = getattr(objects, data['type'])
-            return object_schema.SpeckleObject.parse_obj(data)
-        except:
-            raise, 'Could not parse object :('
+
+    def dict(self):
+        json_string = json.dumps(super(SpeckleObject, self).dict()['properties'])
+
+        self.geometryHash = hashlib.md5(
+            json_string.encode('utf-8')).hexdigest()
+
+        self.hash = hashlib.md5('{}.{}'.format(self.type, json_string).encode('utf-8')).hexdigest()
+
+        return super(SpeckleObject, self).dict()
 
 
 class Resource(ResourceBase):
